+++ src/App.jsx
@@ -8,7 +8,7 @@ import SchoolNavPage from './pages/Overview/SchoolNav/SchoolNav';
 import ListingPage from './pages/Sell/ListingPage';
 import ConfirmationPage from './pages/Sell/ConfirmationPage';
 import { SellProvider } from './contexts/SellContext';
-import NavigationBar from './components/ExternalNavigationBar/ExternalNavBar';
+import NavigationBar from './components/NavigationBar/ExternalNavigationBar/ExternalNavBar';
 import ViewListingPage from './pages/Listing/ViewListingPage';
 import { Providers } from './contexts/Providers';
 import CreateAccount from './pages/CreateAccount';
@@ -18,7 +18,6 @@ import PageWrapper from './components/PageWrapper';
 function App() {
   return (
     <Router>
-
       <NavigationBar />
       <Routes>
         <Route path="/Home" element={<HomePage />}>
