+++ src/components/NavigationBar/InternalNavigationBar/InternalNavBar.jsx
@@ -1,9 +1,10 @@
 import { Link } from 'react-router-dom';
 import './NavigationBar.css';
-import logo from '../../images/logo2.png';
-import NavigationBarItem from './NavigationBarDef';
+import logo from '../../../images/logo2.png';
+import NavigationBarItem from '../NavigationBarDef';
 
 function NavigationBar() {
+  // **** @todo Fix the linking of the nav bar items ***
   return (
     <div className="NavigationBarContainer">
       <div
@@ -18,21 +19,28 @@ function NavigationBar() {
       </div>
       <NavigationBarItem
         float="Left"
-        text="Get started"
-        link="/Home/SchoolNav"
+        text="My account"
+        link="/profile"
         /* id='nav-your-listings' */
       />
       <NavigationBarItem
         float="Left"
-        text="How it works"
-        link="/Home/HowTo"
+        text="Sell"
+        link="/sell"
         /* class='NoDisplayAtPhone600'
         id='nav-about' */
       />
       <NavigationBarItem
         float="Left"
+        text="Buy"
+        link="/buy"
+        /* class='NoDisplayAtPhone480'
+        id='nav-members' */
+      />
+      <NavigationBarItem
+        float="Left"
         text="Home"
-        link="/Home"
+        link="/home"
         /* class='NoDisplayAtPhone480'
         id='nav-members' */
       />
